Add Cypress configurations for E2E tests
Enhanced the configurations of Cypress in its config file. Added 'pageLoadTimeout' to set a 5-second load time for pages, 'baseUrl' to define testing on local server, and 'specPattern' to locate test specs efficiently. This was necessary to establish basic testing environment and make E2E testing more effective and accurate.